refactor(cli): split migrate, unwind the complexity band-aids
`migrate/handler.ts` was 999 lines, one from the repo's 1000-line threshold, so
the next subcommand would have tipped it. Its four `dispatch*` shells and the
cerebro `execute` they serve move to `migrate/dispatch.ts` (162 lines), leaving
`handler.ts` (848) the migration operations themselves — which is also the half an
embedder imports, since `runMigrateGenerateCommand` is re-exported from the
package root. The dependency runs one way: dispatch reads handler, and the
command's `loader` resolves `execute` from dispatch.

The dispatch context also drops `format`: nothing downstream reads it now that
`defineHandler` writes the document.

Two extractions made to satisfy `sonarjs/cognitive-complexity` moved complexity
into parameter bags instead of removing it, and both are unwound:

- `finishExport` threaded six fields out of a scope only to reassemble them, call
  a pre-existing helper and log one line. Inlined at its one call site.
- `emitImportReport` copied six fields out of `outcome` into a literal with the
  same six names. Inlined; the assembly happens where the values are.

Inlining `finishExport` put `runExportCommand` one point over the budget, so the
complexity is actually reduced rather than re-hidden: the target-and-credential
preflight becomes `resolveExportRequest`, which is one cohesive step with one
return value — and the shape `import` has always used for the same job.

`api-snapshots/cli.api.md` is regenerated here for the whole branch: the narrowed
`format` type and the `CommandResult` envelope the two earlier commits introduced
both reach the published surface.
